add validation for migration

+++ database/migrations/2024_10_20_074944_add_state_status_to_posts_table.php
@@ -12,10 +12,14 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('posts', function (Blueprint $table) {
-            $table->unsignedBigInteger('state_id')->nullable()->default(1)->after('content');
-            $table->foreign('state_id')->references('id')->on('states')->onDelete('set null');
+            if (!Schema::hasColumn('posts', 'state_id')) {
+                $table->unsignedBigInteger('state_id')->nullable()->default(1)->after('content');
+                $table->foreign('state_id')->references('id')->on('states')->onDelete('set null');
+            }
 
-            $table->enum('status', ['hidden', 'public'])->default('hidden')->after('state_id');
+            if (!Schema::hasColumn('posts', 'status')) {
+                $table->enum('status', ['hidden', 'public'])->default('hidden')->after('state_id');
+            }
         });
     }
